Minor fixes, Q7 moved to HA2 Dear students,
Q7 in HA1 deals with a simple experiment to estimate the value function, a concept that is not taught yet. Although the assignment text makes it self sufficient, to avoid potential ambiguities, I decided to move it to HA2, which further gives me the flexibility to make its part (ii) non-optional.

Sorry for the inconvenience.

Best,
Sadegh